feat: refactor (PR review tips)
